# Change IntMap.lookup and add new IntMap.query function