- `IntMap.lookup` no longer checks for short circuit failure.
  - Add a new function `IntMap.query` with the old fast-fail behaviour.
  
Addresses https://github.com/haskell/containers/issues/794