@Boarders 2021-09-09
These are the benchmarks I get on my machine (lookup now doesn't fast fail and query is the old behaviour):
```terminal
benchmarked query_hits
time                 276.0 μs   (266.1 μs .. 284.9 μs)
                     0.996 R²   (0.994 R² .. 0.999 R²)
mean                 261.0 μs   (259.0 μs .. 263.7 μs)
std dev              8.073 μs   (5.881 μs .. 11.27 μs)
variance introduced by outliers: 13% (moderately inflated)

benchmarked query_half
time                 135.6 μs   (132.5 μs .. 139.1 μs)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 135.9 μs   (134.9 μs .. 137.1 μs)
std dev              3.739 μs   (3.134 μs .. 4.913 μs)
variance introduced by outliers: 11% (moderately inflated)

benchmarked query_most
time                 248.7 μs   (243.7 μs .. 254.8 μs)
                     0.993 R²   (0.988 R² .. 0.997 R²)
mean                 245.7 μs   (242.5 μs .. 250.9 μs)
std dev              13.82 μs   (9.804 μs .. 21.46 μs)
variance introduced by outliers: 33% (moderately inflated)

benchmarked query_misses
time                 252.6 μs   (241.0 μs .. 261.1 μs)
                     0.983 R²   (0.971 R² .. 0.992 R²)
mean                 309.3 μs   (298.5 μs .. 319.2 μs)
std dev              37.71 μs   (32.95 μs .. 43.33 μs)
variance introduced by outliers: 71% (severely inflated)

benchmarked query_mixed
time                 35.88 μs   (35.03 μs .. 36.58 μs)
                     0.997 R²   (0.996 R² .. 0.999 R²)
mean                 35.55 μs   (35.31 μs .. 35.78 μs)
std dev              809.9 ns   (685.5 ns .. 1.008 μs)

benchmarked lookup_hits
time                 231.5 μs   (227.3 μs .. 237.0 μs)
                     0.996 R²   (0.993 R² .. 0.998 R²)
mean                 240.3 μs   (238.3 μs .. 243.1 μs)
std dev              8.042 μs   (6.614 μs .. 10.46 μs)
variance introduced by outliers: 15% (moderately inflated)

benchmarked lookup_half
time                 236.0 μs   (227.6 μs .. 251.5 μs)
                     0.985 R²   (0.976 R² .. 0.993 R²)
mean                 232.6 μs   (229.1 μs .. 237.5 μs)
std dev              13.75 μs   (10.51 μs .. 17.64 μs)
variance introduced by outliers: 37% (moderately inflated)

benchmarked lookup_most
time                 220.1 μs   (217.9 μs .. 223.4 μs)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 223.8 μs   (222.0 μs .. 225.7 μs)
std dev              6.178 μs   (5.237 μs .. 7.243 μs)
variance introduced by outliers: 11% (moderately inflated)

benchmarked lookup_misses
time                 231.7 μs   (227.4 μs .. 239.1 μs)
                     0.996 R²   (0.991 R² .. 0.999 R²)
mean                 229.9 μs   (228.0 μs .. 232.9 μs)
std dev              7.857 μs   (6.027 μs .. 10.50 μs)
variance introduced by outliers: 16% (moderately inflated)

benchmarked lookup_mixed
time                 20.19 μs   (19.37 μs .. 20.97 μs)
                     0.994 R²   (0.991 R² .. 0.998 R²)
mean                 20.56 μs   (20.31 μs .. 20.81 μs)
std dev              853.2 ns   (692.1 ns .. 1.152 μs)
variance introduced by outliers: 22% (moderately inflated)
```