@treeowl 2021-09-10
And add that it's otherwise slower?

On Thu, Sep 9, 2021, 9:57 PM konsumlamm ***@***.***> wrote:

> ***@***.**** requested changes on this pull request.
> ------------------------------
>
> In containers/src/Data/IntMap/Internal.hs
> <https://github.com/haskell/containers/pull/800#discussion_r705839398>:
>
> > +-- | /O(min(n,W))/. Query has identical behaviour to 'Data.IntMap.Internal.lookup' but
>
> +-- will fail faster in the case that the key does not share  is not present.
>
>
> ⬇️ Suggested change
>
> --- | /O(min(n,W))/. Query has identical behaviour to 'Data.IntMap.Internal.lookup' but
>
> --- will fail faster in the case that the key does not share  is not present.
>
> +-- | /O(min(n,W))/. 'query' has identical behaviour to 'lookup', but
>
> +-- will fail faster in the case that the key is not present.
>
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/haskell/containers/pull/800#pullrequestreview-750984248>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAOOF7N4N7UNI2X7OEDW3ULUBFQY3ANCNFSM5DXX5C6Q>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>
>