@sjakobi 2021-09-15
Apologies for the late response!

> * `IntMap.lookup` no longer checks for short circuit failure.

Frankly, I don't think this is a good idea. We don't know whether any users are relying on `lookup` to fail fast. `containers` is used about everywhere, including packages that don't have active maintainers. I don't think we should release a potentially large performance regression to the ecosystem.

> * Add a new function `IntMap.query` with the old fast-fail behaviour.

I don't think this is good naming. Imagine you're a Haskell newbie, starting to use `Data.IntMap`. How do you remember which of `lookup` and `query` have the desired fail-fast or fail-slow behaviour?

I think it would be better to add a more explicit name, such as `lookupFailSlow` (or `lookupFailFast`).