@treeowl 2021-09-15
I disagree with you on this one, @sjakobi. While it's true we can't change anything without regressing some things, I think in the vast majority of practical cases, lookups are likely to succeed and we should do what we can to make that the fast path. Change logs, announcements, and documentation should prompt anyone relying on short-circuiting to switch functions.