@mitchellwrosen 2021-09-15
Furthermore, from the discussion on the original issue (see https://github.com/haskell/containers/issues/794#issuecomment-916245832 for example), it sounds like it's not as simple as, "if you expect [some vague threshold percentage] of misses, use the fail-fast variant"!