@Boarders 2021-09-15
@mitchellwrosen That is correct (and the above benchmarks show it), the `nomatch` function relies on the prefix stored in the tree to differ before the switching bit of what we are looking up. That means it will be good at short circuiting keys that are far away from those in the map but not those that are close (where closeness is in terms of shared bit prefix). I don't think this is actually a common occurrence and so I don't think the short-circuit default is good.

@sjakobi I understand where you are coming from, why don't I try and see if I can test this change on something that makes large use of intmap (I wanted to try ghc) and see how it goes? I think for smaller consumers of intmap it probably won't make a big difference and for large users we can make the necessary patches to those packages that use the current behaviour in critical ways. I am open to changing the name but I don't think it should be optimized for newbies per se who should not be caring about such issues as to the internals of how intmap works.