@nomeata 2021-09-16
I am on board with @treeowl here: Most changes will benefit some users, and regress some others. My intuition is that this will benefit most users. Definitely those where all lookups succeed, which might actually be a common case (think of the environment in a compiler, all variables will be present in the map). But likely also those users that have failing lookups, but where the looked up keys and the present values are nicely mixed, so the early failure isn’t very early, and the benefit of fewer branches may more. 

But in the end, only benchmarks can tell. Looking forward to @Boarders’s investigation here (although good luck getting statistically significant results here…)