@Boarders 2021-09-16
How about `findWithDefault`?