@Boarders 2021-09-16
Granted, I have no idea what I am doing, but here is what I got from trying to measure this change on GHC building cabal using perf:

original:
```
   777,509,663,120      instructions              #    1.26  insn per cycle         
   617,691,499,581      cycles                                                      

     143.578308984 seconds time elapsed

     142.807996000 seconds user
       1.146751000 seconds sys
```

with this change:
```
   776,453,461,794      instructions              #    1.32  insn per cycle         
   590,402,339,813      cycles                                                      

     134.333660622 seconds time elapsed

     133.517496000 seconds user
       1.228369000 seconds sys
  ```
  
  Doesn't look like there is much in it and I don't really know how variable these sort of measurements are, but definitely doesn't look like any sort of serious regression.